Use new GLSPClient protocl in standalone context
- Refactor Jsonrpc-based protocol implementation into subdirectory
- Replace GLSPWebsocketDiagramServer with GLSPDiagramServer. The GlspDiagramServer uses the newly introduced GLSPClient protocol.
- Restructure features/select and features/validate to get rid of circular dependency warning
