+++ src/features/select/selection-service.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2019 EclipseSource and others.
+ * Copyright (c) 2019-2020 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,13 +14,29 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 import { inject, injectable, multiInject, optional } from "inversify";
-import { Action, ILogger, isSelected, SModelElement, SModelRoot, TYPES } from "sprotty";
+import {
+    Action,
+    Command,
+    CommandExecutionContext,
+    ILogger,
+    isSelectable,
+    isSelected,
+    SChildElement,
+    SelectAction,
+    SelectAllAction,
+    SelectAllCommand as SprottySelectAllCommand,
+    SelectCommand as SprottySelectCommand,
+    SModelElement,
+    SModelRoot,
+    SParentElement,
+    TYPES
+} from "sprotty";
 
 import { SModelRootListener } from "../../base/model/update-model-command";
 import { GLSP_TYPES } from "../../base/types";
 import { distinctAdd, remove } from "../../utils/array-utils";
 import { IFeedbackActionDispatcher } from "../tool-feedback/feedback-action-dispatcher";
-import { SelectFeedbackAction } from "./action-definitions";
+import { SelectFeedbackAction } from "./select-feedback-action";
 
 export interface SelectionListener {
     selectionChanged(root: Readonly<SModelRoot>, selectedElements: string[]): void;
@@ -130,5 +146,95 @@ export class SelectionService implements SModelRootListener {
 
     isMultiSelection(): boolean {
         return this.selectedElementIDs.size > 1;
+    }
+}
+
+
+@injectable()
+export class SelectCommand extends Command {
+    static readonly KIND = SprottySelectCommand.KIND;
+
+    protected selected: SModelElement[] = [];
+    protected deselected: SModelElement[] = [];
+
+    constructor(@inject(TYPES.Action) public action: SelectAction, @inject(GLSP_TYPES.SelectionService) public selectionService: SelectionService) {
+        super();
+    }
+
+    execute(context: CommandExecutionContext): SModelRoot {
+        const model = context.root;
+        this.action.selectedElementsIDs.forEach(id => {
+            const element = model.index.getById(id);
+            if (element instanceof SChildElement && isSelectable(element)) {
+                this.selected.push(element);
+            }
+        });
+        this.action.deselectedElementsIDs.forEach(id => {
+            const element = model.index.getById(id);
+            if (element instanceof SChildElement && isSelectable(element)) {
+                this.deselected.push(element);
+            }
+        });
+        return this.redo(context);
+    }
+
+    undo(context: CommandExecutionContext): SModelRoot {
+        const select = this.deselected.map(element => element.id);
+        const deselect = this.selected.map(element => element.id);
+        this.selectionService.updateSelection(context.root, select, deselect);
+        return context.root;
+    }
+
+    redo(context: CommandExecutionContext): SModelRoot {
+        const select = this.selected.map(element => element.id);
+        const deselect = this.deselected.map(element => element.id);
+        this.selectionService.updateSelection(context.root, select, deselect);
+        return context.root;
+    }
+}
+
+@injectable()
+export class SelectAllCommand extends Command {
+    static readonly KIND = SprottySelectAllCommand.KIND;
+    protected previousSelection: Map<string, boolean> = new Map<string, boolean>();
+
+    constructor(@inject(TYPES.Action) public action: SelectAllAction, @inject(GLSP_TYPES.SelectionService) public selectionService: SelectionService) {
+        super();
+    }
+
+    execute(context: CommandExecutionContext): SModelRoot {
+        return this.redo(context);
+    }
+
+    undo(context: CommandExecutionContext): SModelRoot {
+        const index = context.root.index;
+        for (const previousState of this.previousSelection) {
+            const element = index.getById(previousState[0]);
+            if (element !== undefined && isSelectable(element)) {
+                element.selected = previousState[1];
+            }
+        }
+        return context.root;
+    }
+
+    redo(context: CommandExecutionContext): SModelRoot {
+        const selectables: string[] = [];
+        this.selectAll(context.root, this.action.select, selectables);
+        if (this.action.select) {
+            this.selectionService.updateSelection(context.root, selectables, []);
+        } else {
+            this.selectionService.updateSelection(context.root, [], selectables);
+        }
+        return context.root;
+    }
+
+    protected selectAll(element: SParentElement, newState: boolean, selected: string[]): void {
+        if (isSelectable(element)) {
+            this.previousSelection.set(element.id, element.selected);
+            selected.push(element.id);
+        }
+        for (const child of element.children) {
+            this.selectAll(child, newState, selected);
+        }
     }
 }
